+++ App CODE/App/Screens/Volunteer/Certificates.js
@@ -1,173 +1,89 @@
-import React from "react";
-import { StyleSheet, Text, View, Button, TouchableOpacity } from "react-native";
-import { useState } from "react";
-import { printToFileAsync } from "expo-print";
-import { shareAsync } from "expo-sharing";
-import Icon from "react-native-vector-icons/FontAwesome";
+import React, { useRef, useState } from 'react';
+import { View, Text, StyleSheet, Button, WebView } from 'react-native';
+import * as Print from 'expo-print';
+import * as Sharing from 'expo-sharing';
 
-export default function App() {
-  const volunteerName = "Mahmoud"; // Replace with the actual name variable
+const PdfScreen = () => {
+    const [pdfUrl, setPdfUrl] = useState('');
+    const pdfRef = useRef(null);
 
-  const htmlCertificate = `
-  <html>
-    <head>
-      <style>
-        body {
-          font-family: Arial, sans-serif;
-          background-color: #f2f2f2;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-          min-height: 100vh;
+    const generatePdf = async () => {
+        const html = `
+      <html>
+        <body>
+          <h1>My Dynamic PDF</h1>
+          <p>This PDF was generated dynamically using React Native and Expo.</p>
+          <p>Here's some more text to fill out the PDF.</p>
+        </body>
+      </html>
+    `;
+
+        const options = {
+            html,
+            base64: true,
+        };
+
+        const pdf = await Print.printToFileAsync(options);
+        setPdfUrl(pdf.uri);
+        pdfRef.current = pdf;
+    };
+
+    const handleDownload = async () => {
+        if (!pdfRef.current) {
+            return;
         }
-        .certificate {
-          border: 2px solid #333333;
-          padding: 40px;
-          background-color: #ffffff;
-          text-align: center;
-          max-width: 600px;
-        }
-        .certificate h1 {
-          color: #333333;
-          font-size: 28px;
-          margin-bottom: 20px;
-        }
-        .certificate p {
-          color: #666666;
-          font-size: 18px;
-          line-height: 1.5;
-          margin-bottom: 10px;
-        }
-        .highlight {
-          color: #ff0000;
-          font-weight: bold;
-        }
-      </style>
-    </head>
-    <body>
-      <div class="certificate">
-        <h1>Certificate of Volunteerism</h1>
-        <p>This is to certify that</p>
-        <h2>${volunteerName}</h2>
-        <p>has dedicated their time and efforts as a volunteer in</p>
-        <h3>לאורו נלך</h3>
-      </div>
-    </body>
-  </html>
-`;
 
+        const downloadUrl = pdfRef.current.uri;
+        const filename = 'my-pdf.pdf';
 
-  const totalHours = 10; 
+        await Sharing.shareAsync(downloadUrl, {
+            mimeType: 'application/pdf',
+            dialogTitle: 'Download PDF',
+            UTI: 'com.adobe.pdf',
+            filename,
+        });
+    };
 
-  const cssStyles = `
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      background-color: #f2f2f2;
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      min-height: 100vh;
-    }
-    .certificate {
-      border: 2px solid #333333;
-      padding: 40px;
-      background-color: #ffffff;
-      text-align: center;
-      max-width: 600px;
-    }
-    .certificate h1 {
-      color: #333333;
-      font-size: 28px;
-      margin-bottom: 20px;
-    }
-    .certificate p {
-      color: #666666;
-      font-size: 18px;
-      line-height: 1.5;
-      margin-bottom: 10px;
-    }
-    .highlight {
-      color: #ff0000;
-      font-weight: bold;
-    }
-  </style>
-`;
-
-  const currentDate = new Date().toLocaleDateString("en-US", {
-    year: "numeric",
-    month: "long",
-    day: "numeric",
-  });
-
-  const htmlHoursCertificate = `
-  <html>
-    <head>
-      ${cssStyles}
-    </head>
-    <body>
-      <div class="certificate">
-        <h1>Certificate of Completion</h1>
-        <p>This is to certify that</p>
-        <h2>Your Name</h2>
-        <p>has successfully completed</p>
-        <h3>${totalHours} Hours of Training</h3>
-        <p>Date: ${currentDate}</p>
-      </div>
-    </body>
-  </html>
-`;
-
-  let pdfCertificate = async () => {
-    const file = await printToFileAsync({
-      html: htmlCertificate,
-      base64: false,
-    });
-    await shareAsync(file.uri);
-  };
-
-  let pdfHoursCertificate = async () => {
-    const file = await printToFileAsync({
-      html: htmlHoursCertificate,
-      base64: false,
-    });
-    await shareAsync(file.uri);
-  };
-
-  return (
-    <View style={styles.container}>
-      <TouchableOpacity style={styles.button} onPress={pdfCertificate}>
-        <Icon name="users" size={100} color="black" />
-        <Text style={styles.buttonText}>Certificate</Text>
-      </TouchableOpacity>
-      <TouchableOpacity style={styles.button} onPress={pdfHoursCertificate}>
-        <Icon name="clock-o" size={100} color="black" />
-        <Text style={styles.buttonText}>Hours Certificate</Text>
-      </TouchableOpacity>
-    </View>
-  );
-}
+    return (
+        <View style={styles.container}>
+            {pdfUrl ? (
+                <WebView source={{ uri: pdfUrl }} style={styles.pdfView} />
+            ) : (
+                <View style={styles.emptyView}>
+                    <Text style={styles.emptyText}>No PDF generated yet.</Text>
+                </View>
+            )}
+            <View style={styles.buttonContainer}>
+                <Button title="Generate PDF" onPress={generatePdf} />
+                <Button title="Download PDF" onPress={handleDownload} />
+            </View>
+        </View>
+    );
+};
 
 const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: "#fff",
-    alignItems: "center",
-    justifyContent: "center",
-  },
-  button: {
-    alignItems: "center",
-    justifyContent: "center",
-    backgroundColor: "#D1D5DB",
-    borderRadius: 10,
-    padding: 10,
-    marginVertical: 10,
-    width: 200,
-    height: 200,
-  },
-  buttonText: {
-    fontSize: 16,
-    fontWeight: "bold",
-    marginTop: 10,
-  },
+    container: {
+        flex: 1,
+        padding: 16,
+    },
+    pdfView: {
+        flex: 1,
+    },
+    emptyView: {
+        flex: 1,
+        alignItems: 'center',
+        justifyContent: 'center',
+    },
+    emptyText: {
+        fontSize: 20,
+        fontWeight: 'bold',
+        textAlign: 'center',
+    },
+    buttonContainer: {
+        flexDirection: 'row',
+        justifyContent: 'space-around',
+        marginVertical: 16,
+    },
 });
+
+export default PdfScreen;
