+++ App CODE/App/Screens/WelcomeScreen.js
@@ -94,6 +94,7 @@ const styles = StyleSheet.create({
     justifyContent: 'center',
     alignItems: 'center',
     backgroundColor: '#fff',
+    marginBottom: 100,
 
   },
   logo: {
